+++ ci-test.sh
@@ -29,7 +29,7 @@ else
     # Only testing it for one configuration to avoid running the same tests twice
     echo "Build wasm32 for all examples"
 
-    ./examples/build_all_docker.sh
+    ./examples/build_all_docker.sh --check
     echo "Testing all examples"
     ./examples/test_all.sh
     ./examples/size_all.sh