+++ near-sdk-macros/src/core_impl/abi/mod.rs
@@ -1,3 +1,9 @@
-pub mod abi_generator;
-pub mod abi_visitor;
-mod doc;
+#[cfg(feature = "__abi-embed")]
+mod abi_embed;
+#[cfg(feature = "__abi-embed")]
+pub use abi_embed::embed;
+
+#[cfg(feature = "__abi-generate")]
+mod abi_generator;
+#[cfg(feature = "__abi-generate")]
+pub use abi_generator::generate;