+++ near-sdk-macros/src/core_impl/mod.rs
@@ -1,11 +1,9 @@
-#[cfg(feature = "abi")]
-mod abi;
+#[cfg(any(feature = "__abi-embed", feature = "__abi-generate"))]
+pub(crate) mod abi;
 mod code_generator;
 mod info_extractor;
 mod metadata;
 mod utils;
-#[cfg(feature = "abi")]
-pub use abi::abi_visitor::AbiVisitor;
 pub use code_generator::*;
 pub use info_extractor::*;
 pub use metadata::metadata_visitor::MetadataVisitor;