+++ near-sdk/src/store/lazy_option/impls.rs
@@ -39,7 +39,10 @@ where
         if cfg!(feature = "expensive-debug") {
             self.get().fmt(f)
         } else {
-            f.debug_struct("LazyOption").field("storage_key", &self.storage_key).finish()
+            f.debug_struct("LazyOption")
+                .field("storage_key", &self.prefix)
+                .field("cache", &self.cache.get())
+                .finish()
         }
     }
 }